+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/Raytracing/HDRaytracingManager.cs
@@ -331,6 +331,14 @@ namespace UnityEngine.Rendering.HighDefinition
             return (!materialIsOnlyTransparent && hasTransparentSubMaterial) ? AccelerationStructureStatus.TransparencyIssue : AccelerationStructureStatus.Added;
         }
 
+        // This function mimics the LOD calculation that is used for the rasterization
+        // THe idea is to have rasterization and ray tracing perfectly match.
+        internal float CalculateCameraDistance(float fieldOfView, float worldSpaceSize, float relativeScreenHeight)
+        {
+            float halfAngle = Mathf.Tan(Mathf.Deg2Rad * fieldOfView * 0.5F);
+            return worldSpaceSize / (2.0f * relativeScreenHeight * halfAngle);
+        }
+
         internal void BuildRayTracingAccelerationStructure(HDCamera hdCamera)
         {
             // Clear all the per frame-data
@@ -470,13 +478,24 @@ namespace UnityEngine.Rendering.HighDefinition
                 // Grab the current LOD group
                 LODGroup lodGroup = lodGroupArray[i];
 
+                // Compute the distance between the LOD group and the camera
+                float lodGroupDistance = (hdCamera.camera.transform.position - lodGroup.transform.position).magnitude;
+
+                // Flag to track if the LODGroup has been represented
+                bool lodGroupProcessed = false;
+
                 // Get the set of LODs
                 LOD[] lodArray = lodGroup.GetLODs();
                 for (int lodIdx = 0; lodIdx < lodArray.Length; ++lodIdx)
                 {
+                    // Grab the current LOD
                     LOD currentLOD = lodArray[lodIdx];
-                    // We only want to push to the acceleration structure the lod0, we do not have defined way to select the right LOD at the moment
-                    if (lodIdx == 0)
+
+                    // Compute the distance until which this LOD may be used
+                    float maxLODUsageDistance = CalculateCameraDistance(hdCamera.camera.fieldOfView, lodGroup.size, currentLOD.screenRelativeTransitionHeight);
+
+                    // If the LOD should be the used one
+                    if (!lodGroupProcessed && lodGroupDistance < maxLODUsageDistance)
                     {
                         for (int rendererIdx = 0; rendererIdx < currentLOD.renderers.Length; ++rendererIdx)
                         {
@@ -492,6 +511,9 @@ namespace UnityEngine.Rendering.HighDefinition
                                 recursiveSettings.enable.value, recursiveSettings.layerMask.value,
                                 pathTracingSettings.enable.value, pathTracingSettings.layerMask.value);
                         }
+
+                        // Flag the LOD group as processed
+                        lodGroupProcessed = true;
                     }
 
                     // Add them to the processed set so that they are not taken into account when processing all the renderers